pyproject.toml
[tool.poetry]
name = "vinbot-trading"
version = "0.1.0"
description = "Algorithmic trading bot for Binance with risk management and notifications"
authors = ["Kapoc <kapoc@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "0.110.0"
uvicorn = "0.27.1"
httpx = "0.27.0"
pydantic-settings = "2.2.1"
websockets = "12.0"
python-dotenv = "1.0.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "1.7.4"}
aiosqlite = "^0.20.0"
prometheus-fastapi-instrumentator = "^7.1.0"
opentelemetry-api = "^1.40.0"
opentelemetry-sdk = "^1.40.0"
opentelemetry-instrumentation-fastapi = "^0.61b0"
opentelemetry-exporter-otlp-proto-grpc = "^1.40.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
